+++ lending/src/codatlending/banking.py
@@ -5,7 +5,7 @@ from codatlending import utils
 from codatlending.models import errors, operations, shared
 from typing import Optional
 
-class CashFlow:
+class Banking:
     r"""Retrieve banking data from linked bank accounts."""
     sdk_configuration: SDKConfiguration
 
@@ -148,8 +148,8 @@ class CashFlow:
         return res
 
     
-    def get_enhanced_cash_flow_transactions(self, request: operations.GetEnhancedCashFlowTransactionsRequest, retries: Optional[utils.RetryConfig] = None) -> operations.GetEnhancedCashFlowTransactionsResponse:
-        r"""Get enhanced cash flow report
+    def get_categorized_bank_statement(self, request: operations.GetCategorizedBankStatementRequest, retries: Optional[utils.RetryConfig] = None) -> operations.GetCategorizedBankStatementResponse:
+        r"""Get categorized bank statement
         > **Categorization engine**
         > 
         > The categorization engine uses machine learning and has been fully trained against Plaid and TrueLayer banking data sources. It is not fully trained against the Basiq banking data source.
@@ -158,9 +158,9 @@ class CashFlow:
         """
         base_url = utils.template_url(*self.sdk_configuration.get_server_details())
         
-        url = utils.generate_url(operations.GetEnhancedCashFlowTransactionsRequest, base_url, '/companies/{companyId}/reports/enhancedCashFlow/transactions', request)
+        url = utils.generate_url(operations.GetCategorizedBankStatementRequest, base_url, '/companies/{companyId}/reports/enhancedCashFlow/transactions', request)
         headers = {}
-        query_params = utils.get_query_params(operations.GetEnhancedCashFlowTransactionsRequest, request)
+        query_params = utils.get_query_params(operations.GetCategorizedBankStatementRequest, request)
         headers['Accept'] = 'application/json'
         headers['user-agent'] = f'speakeasy-sdk/{self.sdk_configuration.language} {self.sdk_configuration.sdk_version} {self.sdk_configuration.gen_version} {self.sdk_configuration.openapi_doc_version}'
         
@@ -180,7 +180,7 @@ class CashFlow:
         ]))
         content_type = http_res.headers.get('Content-Type')
 
-        res = operations.GetEnhancedCashFlowTransactionsResponse(status_code=http_res.status_code, content_type=content_type, raw_response=http_res)
+        res = operations.GetCategorizedBankStatementResponse(status_code=http_res.status_code, content_type=content_type, raw_response=http_res)
         
         if http_res.status_code == 200:
             if utils.match_content_type(content_type, 'application/json'):