+++ lending/src/codatlending/financial_statements.py
@@ -5,7 +5,7 @@ from codatlending import utils
 from codatlending.models import errors, operations, shared
 from typing import Optional
 
-class Financials:
+class FinancialStatements:
     r"""Financial data and reports from a linked accounting platform."""
     sdk_configuration: SDKConfiguration
 
@@ -156,17 +156,17 @@ class Financials:
         return res
 
     
-    def get_enhanced_balance_sheet_accounts(self, request: operations.GetEnhancedBalanceSheetAccountsRequest, retries: Optional[utils.RetryConfig] = None) -> operations.GetEnhancedBalanceSheetAccountsResponse:
-        r"""Get enhanced balance sheet accounts
+    def get_categorized_balance_sheet(self, request: operations.GetCategorizedBalanceSheetStatementRequest, retries: Optional[utils.RetryConfig] = None) -> operations.GetCategorizedBalanceSheetStatementResponse:
+        r"""Get categorized balance sheet statement
         The *Get enhanced balance sheet accounts* endpoint returns a list of categorized accounts that appear on a company’s Balance Sheet along with a balance per financial statement date.
 
         Codat suggests a category for each account automatically, but you can [change it](https://docs.codat.io/lending/enhanced-financials/overview#categorize-accounts) to a more suitable one.
         """
         base_url = utils.template_url(*self.sdk_configuration.get_server_details())
         
-        url = utils.generate_url(operations.GetEnhancedBalanceSheetAccountsRequest, base_url, '/companies/{companyId}/reports/enhancedBalanceSheet/accounts', request)
+        url = utils.generate_url(operations.GetCategorizedBalanceSheetStatementRequest, base_url, '/companies/{companyId}/reports/enhancedBalanceSheet/accounts', request)
         headers = {}
-        query_params = utils.get_query_params(operations.GetEnhancedBalanceSheetAccountsRequest, request)
+        query_params = utils.get_query_params(operations.GetCategorizedBalanceSheetStatementRequest, request)
         headers['Accept'] = 'application/json'
         headers['user-agent'] = f'speakeasy-sdk/{self.sdk_configuration.language} {self.sdk_configuration.sdk_version} {self.sdk_configuration.gen_version} {self.sdk_configuration.openapi_doc_version}'
         
@@ -186,7 +186,7 @@ class Financials:
         ]))
         content_type = http_res.headers.get('Content-Type')
 
-        res = operations.GetEnhancedBalanceSheetAccountsResponse(status_code=http_res.status_code, content_type=content_type, raw_response=http_res)
+        res = operations.GetCategorizedBalanceSheetStatementResponse(status_code=http_res.status_code, content_type=content_type, raw_response=http_res)
         
         if http_res.status_code == 200:
             if utils.match_content_type(content_type, 'application/json'):
@@ -204,17 +204,17 @@ class Financials:
         return res
 
     
-    def get_enhanced_profit_and_loss_accounts(self, request: operations.GetEnhancedProfitAndLossAccountsRequest, retries: Optional[utils.RetryConfig] = None) -> operations.GetEnhancedProfitAndLossAccountsResponse:
-        r"""Get enhanced profit and loss accounts
+    def get_categorized_profit_and_loss(self, request: operations.GetCategorizedProfitAndLossStatementRequest, retries: Optional[utils.RetryConfig] = None) -> operations.GetCategorizedProfitAndLossStatementResponse:
+        r"""Get categorized profit and loss statement
         The *Get enhanced profit and loss accounts* endpoint returns a list of categorized accounts that appear on a company’s Profit and Loss statement. It also includes a balance as of the financial statement date.
 
         Codat suggests a category for each account automatically, but you can [change it](https://docs.codat.io/lending/enhanced-financials/overview#categorize-accounts) to a more suitable one.
         """
         base_url = utils.template_url(*self.sdk_configuration.get_server_details())
         
-        url = utils.generate_url(operations.GetEnhancedProfitAndLossAccountsRequest, base_url, '/companies/{companyId}/reports/enhancedProfitAndLoss/accounts', request)
+        url = utils.generate_url(operations.GetCategorizedProfitAndLossStatementRequest, base_url, '/companies/{companyId}/reports/enhancedProfitAndLoss/accounts', request)
         headers = {}
-        query_params = utils.get_query_params(operations.GetEnhancedProfitAndLossAccountsRequest, request)
+        query_params = utils.get_query_params(operations.GetCategorizedProfitAndLossStatementRequest, request)
         headers['Accept'] = 'application/json'
         headers['user-agent'] = f'speakeasy-sdk/{self.sdk_configuration.language} {self.sdk_configuration.sdk_version} {self.sdk_configuration.gen_version} {self.sdk_configuration.openapi_doc_version}'
         
@@ -234,7 +234,7 @@ class Financials:
         ]))
         content_type = http_res.headers.get('Content-Type')
 
-        res = operations.GetEnhancedProfitAndLossAccountsResponse(status_code=http_res.status_code, content_type=content_type, raw_response=http_res)
+        res = operations.GetCategorizedProfitAndLossStatementResponse(status_code=http_res.status_code, content_type=content_type, raw_response=http_res)
         
         if http_res.status_code == 200:
             if utils.match_content_type(content_type, 'application/json'):