+++ lending/src/codatlending/sdk.py
@@ -4,14 +4,14 @@ import requests as requests_http
 from .accounting_bank_data import AccountingBankData
 from .accounts_payable import AccountsPayable
 from .accounts_receivable import AccountsReceivable
-from .cash_flow import CashFlow
+from .banking import Banking
 from .companies import Companies
 from .company_info import CompanyInfo
 from .connections import Connections
 from .data_integrity import DataIntegrity
 from .excel_reports import ExcelReports
 from .file_upload import FileUpload
-from .financials import Financials
+from .financial_statements import FinancialStatements
 from .liabilities import Liabilities
 from .manage_data import ManageData
 from .sales import Sales
@@ -21,13 +21,33 @@ from codatlending import utils
 from codatlending.models import shared
 
 class CodatLending:
-    r"""Lending API: An API for uploading and downloading files from 'File Upload' Integrations.
+    r"""Lending API: Our Lending API helps you make smarter credit decisions on small businesses by enabling you to pull your customers' latest data from accounting, banking, and commerce platforms they are already using. It also includes features to help providers verify the accuracy of data and process it more efficiently.
 
-    The Accounting file upload, Banking file upload, and Business documents file upload integrations provide simple file upload functionality.
+    The Lending API is built on top of the latest accounting, commerce, and banking data, providing you with the most important data points you need to get a full picture of SMB creditworthiness and make a comprehensive assessment of your customers.
 
-    [Read more...](https://docs.codat.io/other/file-upload)
+    [Explore product](https://docs.codat.io/bank-feeds-api/overview) | [See OpenAPI spec](https://github.com/codatio/oas)
 
-    [See our OpenAPI spec](https://github.com/codatio/oas)
+    ---
+
+    ## Endpoints
+
+    | Endpoints            | Description                                                                                                |
+    |:---------------------|:-----------------------------------------------------------------------------------------------------------|
+    | Companies            | Create and manage your SMB users' companies.                                                               |
+    | Connections          | Create new and manage existing data connections for a company.                                             |
+    | Company info         | View company profile from the source platform.                                                             |
+    | Accounts payable     | Data from a linked accounting platform representing money the business owes money to its suppliers.        |
+    | Accounts receivable  | Data from a linked accounting platform representing money owed to the business for sold goods or services. |
+    | Transactions         | Data from a linked accounting platform representing transactions.                                          |
+    | Financial statements | Financial data and reports from a linked accounting platform.                                              |
+    | Banking              | Retrieve banking data from linked bank accounts.                                                           |
+    | Sales                | Retrieve standardized sales data from a linked commerce platform.                                          |
+    | Liabilities          | Debt and other liabilities.                                                                                |
+    | Data integrity       | Match mutable accounting data with immutable banking data to increase confidence in financial data.        |
+    | Excel reports        | Download reports in Excel format.                                                                          |
+    | Categories           | Manage Codat's automatic account categorization functionality.                                             |
+    | Manage data          | Control how data is retrieved from an integration.                                                         |
+    | File upload          | Endpoints to manage uploaded files.                                                                        |
     """
     accounting_bank_data: AccountingBankData
     r"""Access bank transactions from an accounting platform."""
@@ -35,7 +55,7 @@ class CodatLending:
     r"""Data from a linked accounting platform representing money the business owes money to its suppliers."""
     accounts_receivable: AccountsReceivable
     r"""Data from a linked accounting platform representing money owed to the business for sold goods or services."""
-    cash_flow: CashFlow
+    banking: Banking
     r"""Retrieve banking data from linked bank accounts."""
     companies: Companies
     r"""Create and manage your Codat companies."""
@@ -49,7 +69,7 @@ class CodatLending:
     r"""Downloadable reports"""
     file_upload: FileUpload
     r"""Endpoints to manage uploaded files."""
-    financials: Financials
+    financial_statements: FinancialStatements
     r"""Financial data and reports from a linked accounting platform."""
     liabilities: Liabilities
     r"""Debt and other liabilities."""
@@ -99,14 +119,14 @@ class CodatLending:
         self.accounting_bank_data = AccountingBankData(self.sdk_configuration)
         self.accounts_payable = AccountsPayable(self.sdk_configuration)
         self.accounts_receivable = AccountsReceivable(self.sdk_configuration)
-        self.cash_flow = CashFlow(self.sdk_configuration)
+        self.banking = Banking(self.sdk_configuration)
         self.companies = Companies(self.sdk_configuration)
         self.company_info = CompanyInfo(self.sdk_configuration)
         self.connections = Connections(self.sdk_configuration)
         self.data_integrity = DataIntegrity(self.sdk_configuration)
         self.excel_reports = ExcelReports(self.sdk_configuration)
         self.file_upload = FileUpload(self.sdk_configuration)
-        self.financials = Financials(self.sdk_configuration)
+        self.financial_statements = FinancialStatements(self.sdk_configuration)
         self.liabilities = Liabilities(self.sdk_configuration)
         self.manage_data = ManageData(self.sdk_configuration)
         self.sales = Sales(self.sdk_configuration)