+++ lending/RELEASES.md
@@ -9,3 +9,13 @@ Based on:
 - [python v0.1.0] lending
 ### Releases
 - [PyPI v0.1.0] https://pypi.org/project/codat-lending/0.1.0 - lending
+
+## 2023-09-07 00:15:51
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Lending.yaml
+- Speakeasy CLI 1.79.0 (2.101.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v0.2.0] lending
+### Releases
+- [PyPI v0.2.0] https://pypi.org/project/codat-lending/0.2.0 - lending